first pass response to review of pull-request and isolating squid from permissions.
